Remove `Config` support from the crate temporarily.
As @ZER0 suggested, it would be better to avoid for now
supporting a configfile + `Config` struct since we still
are on early stages of development. (And even makes sense to
have that from the beggining, it will take us time from
higher priority tasks).

So refactored the crate to remove the configfile support.