Disable autobins flag for crate
By default, Cargo automatically determines the targets to
build based on the layout of the files on the filesystem.
The target configuration tables, such as [lib], [[bin]],
[[test]], [[bench]], or [[example]], can be used to add
additional targets that don't follow the standard directory
layout.

The problem comes when you add the config tables and since
you have a `bin/` folder, cargo decides that all of the files
inside are a binary(package).
Therefore, we were getting errors due to the fact that the files
at the same level to `main.rs` were considered different binaries.