Create an encoding module
We need to encode/decode a lot of GRPC types in order to be
able to propperly working with them.

Therefore, since they are almost always the same ones, a module
for encoding fn's such as `TryFrom` impls has been created.

Added `TryFrom` impl for `BlsScalar` & `JubJubAffine`.
Added a generic fn that tries to get optional parameters
from a tonic `Request` returning an `Err` if any of them
is not present (is `None`).