Add decoding of request parameters
This includes the addition of a mocked BN256Point
type in the rusk library, which is used
as a placeholder for now, to allow for some
progress on the code.

The decoding of the note is currently commented
out, as importing the phoenix-core crate
caused some compilation errors, possibly
due to using an outdated version of dusk-pki.
This should be investigated further.