Optimize & Speed Up Pub Params usage/read
We were previously reading the public parameters from disk
on every call that arrived requiering a proff generation/verification.
It was unoptimal to work on that way since due to the ammount
of checks that are needed to guarantee that the Public Parameters
are correct is massive, it takes like 2min to obtain an instance
of the Public Parameters.

So we've moved the Pub Params read to a lazy_static block which is
executed in the `default` fn call to the Rusk instance that you execute.
This forces to read the pub params just at the time where you
instanciate the Rusk server and therefore no RPC method gets delayed
by this fact.

Closes #112