Implement custom Resolver for Genesis Contracts
We couldn't add the host functions to `canonical`
and therefore now that https://github.com/dusk-network/canonical/pull/49/
landed we should implement our own resolver and load it
during the WASMI initialization.

- Added `RuskExternals` structure to `rusk` lib
including `poseidon_hash` & `signature_verif` functions.
- Added a `MemoryRef` attribute to the `RuskExternals` so
that we can access to the fields that are sent through the
FFI calls.

Closes #141