Implement Proof verification for RuskResolver
In order to be able to implement any contract that contains
functions that need to verify a ZKProof, we need to basically
have a host fn implemented for `RuskExternals` which allows to do so.

This function has been implemented on a way which is callable from any
contract on a generic way.