Transfer Circuits - From public inputs
Closes #225

Currently, the PLONK circuit API relies on a valid instance to
set the correct public inputs positions so it can run proofs.

The host must be agnostic of the circuit instance because the
public inputs exists only on the contract scope. But, it still
need the circuit instance to call the verify methods of the
plonk traits.