rusk-abi: Change `verify_proof` to take `VerifierData`
The verifier key was upgraded to contain also the public input positions
to avoid unnecessary complexity while building the static data of the
used circuits.

This change is transparent to the users of plonk library, so we needed
to remove the public inputs positions as a parameter of the ABI and
treat the verifier agnostically as a vec of u8.

Resolves: #247