Change `verify_proof` to accept verifier data
Resolves #247

To improve the build process of the static data of circuits by avoiding
code complexity, the verifier key was upgraded to contain both the
verifier key and the public inputs positions into a new structure called
verifier data.

This change is transparent to the users of plonk library, so we needed
to remove the public inputs positions as a parameter of the ABI and
treat the verifier data agnostically as a `Vec<u8>`.