code-hasher: Add code_hasher attribute macro impl
We need a way to version our circuits. A first step into
that direction was basically the fact of introducing the CIRCUIT_ID
associated const in the Circuit trait. See: dusk-network/plonk#454

This is the second step which hashes the contents of a code block and
toghether with a provided version and an `IDENT` generates an unique
identifier for the whole block adding the const declaration to it.

- Added test cases suggested by @zer0
- Went for the generic approach suggested by @zer0 in #249 review which
simplifies the implementation in order to be able to complicate less the
things for now.
- The hash includes the versions:
1. Passed by attribute in the macro (is optional).
2. The version of the `code_hasher` crate.

Resolves: #240