code-hasher: Remove CARGO_PKG_VERSION from hash
As @zer0 said, if we link the hash to the crate version, we might have
behaviours that we don't want since every time a new minor version is
out for the macro, we will get a new hash. And for now that's not the
expected behaviour that we're looking for.