Change `verify_proof` to accept `VerifierData`
Plonk 0.7 works with verifier data instead of verifier key so the public
inputs positions can be concatenated to a single slice of bytes and
proofs can be easily verified without any knowledge of its contents.

This is a major change on Rusk structure to accept this new way of
proving routines.

Resolves #247