transfer-circuits: Verifier data optimization
The new 0.7 version of dusk-plonk brings some major optimizations to the
verification process.

Also, a major bug was fixed in version 0.20 of dusk-poseidon that caused
the merkle opening to create different circuit descriptions depending on
the witness input. This led to inconsistencies regarding the verifier
key that must be constant for any given circuit, including the variants
of the execute circuit.

Resolves #262