transfer-circuits: CTCO schnorr payload scheme
The schnorr signature binds the full message of the send to contract
obfuscated calls to its proof.

Besides that, the circuit may take the associated PK privately or
publicly. This should not implicate multiple circuit descriptions for
STCO, so it is constructed with an identity points scheme that takes a
public input stating if the message PK will come from a public input or
witness data.

Resolves #267