transfer-circuits: Change ExecuteCircuit to split
Currently the transfer circuit is a single struct implementation that
will iterate its attributes to produce different circuits.

This doesn't work with the new circuit hasher structure because it uses
the gadget code to generate the identifier of a circuit.

The current implementation will produce different circuits with the same
gadget code depending on the inputs.

This commit splits the `ExecuteCircuit` into different concrete
structures and uses an enum mapping to encapsulate the inputs and
outputs operations.

Resolves #275