transfer-contract: Add `circuits` as default feature
This change was needed because the recovery-tool, it was a kind of
"chicken-egg" situation: the tools was used to create the keys
and deploy the state, but in order to deploy the state, that uses the
genesis contracts, we needed the keys at compile time, because they
are used in the genesis contracts.

However, in order to deploy just the state the circuits are not needed,
only the contract's structure. So with the `circuits` feature we can
avoid to load that part of the code when we're build the tool.

It's not an ideal solution, but it's a working one for the time being.