rusk-recovery: split `state` and `keys` subcommands in two binaries
The state subcommand needs to have the WASM file compiled in order to
deploy the initial state of the network with the genesis contracts.
However, the contracts needs the `keys` in order to be compiled in WASM.
Therefore is not possible to compile one single binary that perform
both operations.
The solution here is having two binaries that still share the same
logic via the same library, but compiles with different requirements
using feature flags.

Resolves: #441

 - Use clap's derive API

Resolves: #436