rusk: reduce number of coinbase transactions to 1
The new coinbase transaction contains both the note for Dusk and the
note for the block generator serialized in sequence. The transaction
hash is calculated using `rusk_abi::Hasher` instead of the significantly
more expensive `dusk_poseidon::sponge::hash`. This is in line with how
normal transfer transactions are hashed.

Resolves: #529