wallet: fix denominations and ask for gas price
The denominations are normalized to always be asked in Dusk from the
user, and converted to μDusk when used internally. Some sane - aka pass
verification - defaults are added.

Gas price is asked from the user as well. The gas price, gas limit and
transfer amount have a lower limit of 1 μDusk, since anything lower is
floored.

Resolves #544