wallet: Invalid config shouldnt fallback
Currently we are using a fallback strategy to pick sequences of
configurations until a valid one is found.

This is misleading in the sense of making the user believe he is working
with a valid configuration, when in fact the wallet is using a different
one - possibly with undesirable parameters.

Resolves #670