Update note picker to always pick the most notes
Performing a transaction usually generates two output notes - one note
intended for the target of the transaction, and a change note for the
transactor. When numerous transactions are performed, the previous
algorithm leads to a proliferation of small change notes in a user's
wallet exacerbating the fragmentation issue, eventually preventing a
user from transacting larger amounts.

This algorithm is an effort to mitigate the fragmentation issue, by
always picking the maximum number of notes available and allowed, and
minimizing usage of large notes. This leads to to the desired asymptotic
behavior of minimizing the number of notes in a user's wallet.

Resolves #55