transfer-circuits: Remove output circuit variants
The execute circuit variants increases the storage requirements of Rusk.
The variants can be reduced, depending only on the inputs.

The only check that is performed with the outputs is the homomorphic
sum, and this is trivial to be mocked via public inputs with zeroed
pedersen commitments.

This PR introduces a circuit padding with these special picked values.

Resolves #693