`stake-contract-types`: initial crate
Introducing stake types reduces the dependencies of the stake contract
itself and allows for a user of said contract to easily structure a
transaction with it.
We also refactor the stake contract to make use of these changes and
introduce a series of queries and (management) transactions, meant to
report on the current state of the contract and modifying it based on
the consensus, respectively.

Co-authored-by: Milosz Muszynski <milosz@dusk.network>
Co-authored-by: OlSe <os-dusk-github@oms.linux-one.org>