rusk: migrate to `piecrust`
Upgrade dependencies and change to make use of the new `piecrust`
virtual machine.

The main change consists in removing the concept of a `RuskState` and
the respective struct. The state is no longer fully manipulable from
the host, and all interactions with it must be done either through
transactions or queries - i.e. by running the VM. The genesis contracts
are tailored to support this behavior. On the same token, `microkelvin`
is no longer used to manage state persistance and this is left fully
to the VM.

This also means the `Rusk` struct can become much simpler and more
useful, in the sense that it is now much easier to simply instantiate it
given the state directory path, and queries being exposed through it,
together with state transitions.

Resolves #759