rusk: handle gas spent returned from TX execution
The transfer contract now returns the gas charged to a transaction, and
we use that as an input for the rewards calculation, as opposed to the
total gas spent in the execution of a contract.

Since the gas spent in the execution of the transfer contract is a
larger number than the gas charged to a transaction - due to the fact
that some code is executed *after* charging - this prevents a small
inflationary effect from happening.

Resolves #790