rusk: update root on every block and upgrade to `dusk-merkle`
The root of the transfer contract tree gets updated after execution of a
block, together with the reward calculation.

Additionally since it is possible for a transaction to require more gas
to execute than is actually charged to it - due to some code having to
be executed *after* the charging code - we introduce "grace gas". This
is an amount of gas approximately equal to the amount spent by the code
after the charging code, allowing for the user to only have to provide a
gas limit similar to what is actually going to be spent.