rusk-abi: upgrade to `piecrust@0.6`
This also entails exposing features appropriately. In this case, we
choose to expose the following features:

- `abi` for crates wishing to write contracts
- `debug` for writing contracts with access to the debug ABI
- `dlmalloc` for including its namesake as an allocator
- `host` for crates wishing to write hosts compatible with the ABI
- `host_debug` for writing hosts with the debug ABI

By default, the `abi` and `dlmalloc` features are turned on. If the user
decides to build without default features, only the types will exposed,
which can sometimes be useful for `no_std` targets that are *not*
contracts.