rusk: initial impl of request-response queries
This is an implementation of a simple request-response protocol for
querying the state of the chain over WebSockets.

When a stream is estabilished, a task is started to handle incoming
requests. Once a request comes in we start a new task. This will report
the result of the chain query which the request represents back to the
stream handling task. This is done to ensure we don't block before
another request can be received.

Resolves #933