rusk: introduce `QueryRaw` trait
The `QueryRaw` trait allows for the WebSocket server to consume a struct
that can be queried in the way a VM can be queried. This allows for
testing the behavior of the server by providing mock implementations
that return some pre-defined data or error.

In the spirit of testing we also introduce the `multiple_queries` test,
showing that the server can be queried multiple times before the client
receives any responses.