rusk: handle HTTP requests for VM execution
Changes the name of the `WsServer` struct to `HttpServer` and makes it
also handle VM executions with the same request type - now renamed to
`ExecutionRequest`.

Since when a response - now renamed to `ExecutionResponse` - is returned
under the context of HTTP the request that it belongs to is clear, the
`request_id` field is made optional in both the request and the
response.