Update workflows to be skipped properly
As stated by the github docs, if a workflow is skipped due to path
filtering, branch filtering or a commit message, then checks associated
with that workflow will remain in a "Pending" state.
A pull request that requires those checks to be successful will be
blocked from merging.
The solution is skipping those workflow without using the `paths`
mechanism, but instead having a job that verifies if something changed
in the interested paths and set a variable, to use on other jobs.