rusk: allow queries in parallel with state transitions
State transitions can now be performed in parallel with queries. This
solves the issue of the consensus being blocked while the state is
queried.

This is achieved by disentangling the `VM` from the current and base
commits in the `Rusk` data structure, as well as using a `RWLock`
instead of a `Mutex`, allowing for write-locking only at the end of a
state transition. This, together with a read-lock and release for when
the tip is used, effectively means that queries are never in contention
with state transitions when they don't need to be.

We also ensure that, on a `finalize`, commits are deleted on a separate
task, such that we don't block execution of the state transition itself,
allowing the node to return from a `finalize` before the deletions are
processed.

Resolves: #970
See-also: #1084