rusk: implement listening to emitted events
We leverage a first draft implementation of RUES to allow websocket
clients to subscribe to events emitted by contracts during a block
transition (AST/FST).

We don't fully implement RUES, since it the we're still missing the
capability of handling dispatched events from the clien, but we do allow
for the basic subscribe/unsubscribe functionality. We also explictly
diverge from the spec in two ways:

- Use the `Sec-WebSocket-Accept` header, instead of `Sec-WebSocket-Key`,
  to generate the session ID. This allows us to stay more in line with
  the way in which RFC6455 handles the websocket session itself.
- We encode the contract's address in HEX, as opposed to base58, since
  we already have used this format in the explorer and the previous API.
  It also has the advantage of being fixed length.

Resolves: #931