transfer: add bridge function
This commit introduces a new bridge functionality to the transfer
contract's state. This function enables any sequencer to register as a
bridge provider upon staking a balance into the bridge. The authorized
sequencer can then facilitate up to the amount of their stake in bridge
operations.

Adopting an optimistic approach, the bridge will process transactions
without immediately verifying L1 associated events. However, it will
keep track of these events and wait for several blocks before
considering a transaction finalized. If a fraud proof is presented,
demonstrating that a bridge operation was carried out without a valid L1
event, the responsible sequencer stands to lose all rewards and staked
amount. In contrast, the individual who exposed the fraudulent activity
will be rewarded with the ill-gotten amount.