execution-core: add Moonlight and rework withdrawals
With this commit we add `MoonlightTransaction`, and rename the existing
transaction to `PhoenixTransaction`, both of which are incorporated into
an enum named `Transaction`.

We also rework the withdrawals system, under the assumption that there
must be a strong assurance that the transacting party own the addresses
being withdraw to. This enables cross-withdrawals from Moonlight to
Phoenix (and vice-versa), as well as withdrawals to the same transaction
model.

The liberty is taken to change the stake data structures slighty, to
improve on clarity, as well as make use of the new withdrawal system.