rusk-abi: add BLS signature cache
Cache definition placed in a different file, and abstracted away with a
macro, allowing one to generate different cache types according to
in-code parameters.

This macro is then leveraged to generate both Plonk proof and BLS
signature caches, which are then subsequently used in our code.

To slightly improve the implementation of both, the hash used as a cache
key is now computed using the argument buffer contents, as opposed to
the deserialized parameters. This allows for a cleaner and more
performant implementation.

See-also: #1984