node: fix "freed while still in use" error
The `tokio::select!` macro seems to have changed in such a way as to not
work well with the syntax we used, and in the process breaking their
crate even if the version doesn't indicate it under semver rules.

This commit addresses this by changing the syntax to something the macro
understands.