consensus: fix verify_att
Fixes a bug in verify_att, which returned the set of all committee
members instead of just the actual voters.

verify_votes is modified to return the set of voters corresponding
to the quorum committee.

verify_step_votes is modified to return a vector of voters instead
of a committee.