execution-core: use BLS single-key scheme
By updating `bls12_381_bls` to `0.4` and changing the API to conform to
it, as well as modifying the use of the functions accordingly, we change
the scheme used for Moonlight and stake to be single-key, as opposed to
multi-signature.

This is not only faster, since it doesn't require key or signature
aggregation, but also makes more semantic sense.