transfer-contract: add `transfer_to_contract` function
The `transfer_to_contract` function can be called by a sender contract
to transfer Dusk to a receiver contract. Once called, a function
specified by the sender is called on the receiver contract to inform it
of the sender's ID and the amount of Dusk sent. The receiver can choose
to accept the transfer, by successfully concluding the execution of the
function called to inform it, or it can panic and effectively reject the
transfer.

The `TransferToContract` is used as an argument to the `transfer_to_contract`,
and is constructed by the contract calling it. `ReceiveFromContract` is the
ata the receiver contract *must* accept for a function to be able to
successfully receive funds from another contract. If a contract wished
to expose one or more of these functions, it is heavily recommended that
they panic if called by anyone else apart from the transfer contract.