execution-core: add `ContractExec::Memo`
Memos are often used by exchanges to use a single account and yet still
be able to differentiate between internal accounts.

In this implementation we reuse the existing `ContractExec` infra and
add a separate variant named `ContractExec::Memo` holding 512 bytes, or
0.5 KiBs, which should be enough to hold any exchange-internal account
as well as any public-key or address necessary.

We also ensure that the enum continues to make sense semantically, since
it now carries more than just interactions with contracts. Renaming from
`ContractExec` to `TransactionData` makes sense since it is now just
"some sort of data" a transaction may carry.

In addition we also rename the `contract_exec` module to data, justified
in exactly the same way as above.

Resolves: #2264