transfer-contract: adjust event emittance
The following events are added, which are emitted when the function with
their same name is successful:

- `mint`
- `withdraw`
- `convert`
- `deposit`
- `transfer_to_contract`
- `transfer_to_account`

Additionally an event is emitted at the end of any transaction
ingestion. This event will either be `phoenix` or `moonlight`, depending
on the model used, and are emitted during the host's call to `refund`.
This is so that the event data can contain both the change and gas
spent.

As a consequence of the above changes, we needed to slightly change the
way in which we insert notes into the tree, and make some changes to the
`transitory` module to support including the change note in the emission
event.

See-also: #2265