node: send local blk if better than remote
If the node receives a block with higher iteration than the local
block, it sends back the local block. The peer is expected to
revert and accept it, unless the block sent to us has been
finalized.

This commit also refactors on_block_event to unify the handling
of remote_blk.height < tip.height and remote_blk.height == tip.height.
These two cases are in fact equivalent and should be treated the same.

The unification of the two cases greatly improves readability and
maintenance. A bug is also fixed, due to which  the node erroneously
switched to OutOfSync if reverting a block lower than the tip.