consensus: fix get_fail_att
Make get_fail_att filter out attestations with no quorum.
Before this change, the function would return an Attestation even
if it didn't reach a quorum. As a result, Fail Attestations from the
network were not stored even if we had a partial attestation.
Now, if we have a partial attestation and we receive a Fail
Attestation from the network, we store it.
This allow the attestation to be included in the failed iterations
list when building a new block.