consensus: skip unnecessary ValidationResult verifications
This commit changes the Ratification collect to only verify the
ValidationResult in the Ratification vote if it has a quorum and
our result is NoQuorum.
If the verification succeeds, we also replace our result with the
received ValidationResult.

We also change the Quorum message to use the local ValidationResult
(possibly updated from the one from a Ratification vote) instead
of the ValidationResult from the last vote we received.